Move CFD-related fucntions to a separate module.
   - The compute follows data related functions are
     required in parfor lowering as well. The PR moves
     the two functions determine_kernel_launch_queue and
     chk_compute_follows_data_compliance out of
     kernel_interface/dispatcher.py into kernel_interface/utils.py
     and makes them public.
   - The Range and NdRange classes are moved to
     kernel_interface/indexers.py
