+++ numba_dpex/tests/kernel_tests/test_kernel_launch_params.py
@@ -12,7 +12,7 @@ from numba_dpex.core.exceptions import (
     InvalidKernelLaunchArgsError,
     UnknownGlobalRangeError,
 )
-from numba_dpex.core.kernel_interface.utils import Range
+from numba_dpex.core.kernel_interface.indexers import Range
 
 
 @dpex.kernel
