+++ numba_dpex/tests/kernel_tests/test_kernel_specialization.py
@@ -11,7 +11,7 @@ from numba_dpex.core.exceptions import (
     InvalidKernelSpecializationError,
     MissingSpecializationError,
 )
-from numba_dpex.core.kernel_interface.utils import Range
+from numba_dpex.core.kernel_interface.indexers import Range
 
 i64arrty = usm_ndarray(ndim=1, dtype=int64, layout="C")
 f32arrty = usm_ndarray(ndim=1, dtype=float32, layout="C")
